created wscreen.go updated readme
